+++ server/src/de/mpg/mpi/inf/d5/yavi/server/resources/SignalResource.java
@@ -22,7 +22,7 @@ public class SignalResource {
   
   @GET
   @JSONP(queryParam = JSONP.DEFAULT_CALLBACK)
-  public List<DateValue> doGet(@QueryParam("pageid") long pageId,
+  public List<DateValue> doGet(@QueryParam("pageid") int pageId,
       @QueryParam("wikipediaid") String wikipediaId,
       @QueryParam("datefrom") String dateFrom,
       @QueryParam("dateto") String dateTo,
