+++ server/src/de/mpg/mpi/inf/d5/yavi/server/tables/AddedOutlinkPK.java
@@ -13,19 +13,19 @@ public class AddedOutlinkPK implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Column(name="page_id")
-	private Long pageId;
+	private Integer pageId;
 
 	@Column(name="day_number")
 	private Integer dayNumber;
 
-	private Long outlink;
+	private Integer outlink;
 
 	public AddedOutlinkPK() {
 	}
-	public Long getPageId() {
+	public Integer getPageId() {
 		return this.pageId;
 	}
-	public void setPageId(Long pageId) {
+	public void setPageId(Integer pageId) {
 		this.pageId = pageId;
 	}
 	public Integer getDayNumber() {
@@ -34,10 +34,10 @@ public class AddedOutlinkPK implements Serializable {
 	public void setDayNumber(Integer dayNumber) {
 		this.dayNumber = dayNumber;
 	}
-	public Long getOutlink() {
+	public Integer getOutlink() {
 		return this.outlink;
 	}
-	public void setOutlink(Long outlink) {
+	public void setOutlink(Integer outlink) {
 		this.outlink = outlink;
 	}
