+++ server/src/de/mpg/mpi/inf/d5/yavi/server/tables/NumberOfAddedOutlinksPK.java
@@ -13,17 +13,17 @@ public class NumberOfAddedOutlinksPK implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Column(name="page_id")
-	private Long pageId;
+	private Integer pageId;
 
 	@Column(name="day_number")
 	private Integer dayNumber;
 
 	public NumberOfAddedOutlinksPK() {
 	}
-	public Long getPageId() {
+	public Integer getPageId() {
 		return this.pageId;
 	}
-	public void setPageId(Long pageId) {
+	public void setPageId(Integer pageId) {
 		this.pageId = pageId;
 	}
 	public Integer getDayNumber() {
