+++ server/src/de/mpg/mpi/inf/d5/yavi/server/tables/NumberOfRevisionsPK.java
@@ -13,17 +13,17 @@ public class NumberOfRevisionsPK implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Column(name="page_id")
-	private Long pageId;
+	private Integer pageId;
 
 	@Column(name="day_number")
 	private Integer dayNumber;
 
 	public NumberOfRevisionsPK() {
 	}
-	public Long getPageId() {
+	public Integer getPageId() {
 		return this.pageId;
 	}
-	public void setPageId(Long pageId) {
+	public void setPageId(Integer pageId) {
 		this.pageId = pageId;
 	}
 	public Integer getDayNumber() {
